# aws/types: grow WriteAtBuffer capacity by 2m when growing to length m