--- a/aws/types.go
+++ b/aws/types.go
@@ -61,6 +61,18 @@ func (r ReaderSeekerCloser) Close() error {
 type WriteAtBuffer struct {
 	buf []byte
 	m   sync.Mutex
+
+	// GrowthCoeff defines the growth rate of the internal buffer. By
+	// default, the growth rate is 1, where expanding the internal
+	// buffer will allocate only enough capacity to fit the new expected
+	// length.
+	GrowthCoeff float64
+}
+
+// NewWriteAtBuffer creates a WriteAtBuffer with an internal buffer
+// provided by buf.
+func NewWriteAtBuffer(buf []byte) *WriteAtBuffer {
+	return &WriteAtBuffer{buf: buf}
 }
 
 // WriteAt writes a slice of bytes to a buffer starting at the position provided
@@ -73,7 +85,10 @@ func (b *WriteAtBuffer) WriteAt(p []byte, pos int64) (n int, err error) {
 	defer b.m.Unlock()
 	if int64(len(b.buf)) < expLen {
 		if int64(cap(b.buf)) < expLen {
-			newBuf := make([]byte, expLen, 2*expLen)
+			if b.GrowthCoeff < 1 {
+				b.GrowthCoeff = 1
+			}
+			newBuf := make([]byte, expLen, int64(b.GrowthCoeff*float64(expLen)))
 			copy(newBuf, b.buf)
 			b.buf = newBuf
 		}