--- a/aws/types_test.go
+++ b/aws/types_test.go
@@ -54,3 +54,22 @@ func BenchmarkWriteAtBufferParallel(b *testing.B) {
 		}
 	})
 }
+
+func BenchmarkWriteAtBufferOrderedWrites(b *testing.B) {
+	// test the performance of a WriteAtBuffer when written in an
+	// ordered fashion. This is similar to the behavior of the
+	// s3.Downloader, since downloads the first chunk of the file, then
+	// the second, and so on.
+	//
+	// This test simulates a 150MB file being written in 30 ordered 5MB chunks.
+	chunk := int64(5e6)
+	max := chunk * 30
+	// we'll write the same 5MB chunk every time
+	tmp := make([]byte, chunk)
+	for i := 0; i < b.N; i++ {
+		buf := &WriteAtBuffer{}
+		for i := int64(0); i < max; i += chunk {
+			buf.WriteAt(tmp, i)
+		}
+	}
+}