@xibz 2016-02-26
@tomwans, thank you for taking the time to do this PR. One suggestion is to add a function pointer to the WriteAtBuffer structure, GrowthFn. This could be then replace this line ```newBuf := make([]byte, expLen, 2*expLen)``` with ```newBuf := make([]byte, expLen, b.GrowthFn(expLen, pLen))```. If the GrowthFn is not set then default to `2 * pLen` to `4 * pLen`. This will allow people to choose how they want to grow. The reason for this is if someone was to pass a buffer that is fairly large in size, 1.5  gigs, this would allocate a buffer of 3 gigs. Does that sound reasonable?