@tomwans 2016-03-24
Apologies for the radio silence here, been out on vacation for a bit.

I didn’t include the ability to change the growth rate here because the WriteAtBuffer is essentially a concurrent dynamic array. I  initially picked 2m because I thought it was ~optimal with respect to the number of times we had to copy old elements after a resize. When the buffer is at its maximum size m, m/2 of the elements would have been copied only once, m/4 only twice, and so on. On average, every element would be copied twice. This would mean that the work to manage the internal buffer is about linear (O(2n) = O(n)) to the number of elements in the array, which I figured was pretty good.

I did some googling to see if there is a good rule of thumb around the growth rate for dynamic arrays, and as you may expect I came across a ton of different answers. I’ve been convinced by https://github.com/facebook/folly/blob/master/folly/docs/FBVector.md that it’s probably best to just go with 1.5m for now.

What do you say to that? If y’all still want the ability to define the growth rate via a passed in func, I’m happy to do that, but just trying to keep things simple here since it seems like 1.5m should be enough for most cases.