@xibz 2016-03-24
@tomwans - No worries! Also, thank you for the research.

Another solution could be to pass your own buffer and we just use that with some growth rate with the limit of the buffer you passed in.
Or pass a growth rate/max size value, where growth rate of zero defaults to one.
What do you think of these? The first would allow for no buffer reallocation and the ceiling would be the size of the buffer. The second would just allow for users to at least specify how fast they want their buffer to grow or cap. What do you think of these?