@tomwans 2016-03-28
Hey @xibz I updated this PR with your suggestions. Let me know if you have any other feedback.